알림 상태에 따라 헤더 알림 표시 변경되도록 수정
